check-gibber-r-multicast-packet: Change part/total/stream_id types for the data packet to it's new types
20071017210637-b58c9-6af11460a5d9d4ce1f486d826a31a84140a2a8c4.gz

+++ tests/check-gibber-r-multicast-packet.c
@@ -53,8 +53,8 @@ START_TEST (test_data_packet) {
   GibberRMulticastPacket *b;
   guint32 sender_id = 1234;
   guint32 packet_id = 1200;
-  guint8 part = 2, total = 3;
-  guint8 stream_id = 56;
+  guint32 part = 2, total = 3;
+  guint16 stream_id = 56;
   guint8 *data;
   gsize len;
   guint8 *pdata;
